docs(Storybook): sort and group adaptive cards stories
